refactor: rename test folder for `TableSchema`
