+++ Runtime/safe-ds/safe_ds/data/__init__.py
@@ -2,4 +2,4 @@ from ._column import Column
 from ._row import Row
 from ._supervised_dataset import SupervisedDataset
 from ._table import Table
-from ._tableSchema import TableSchema
+from ._table_schema import TableSchema
