+++ Runtime/safe-ds/safe_ds/data/_table_schema.py
