+++ Runtime/safe-ds/tests/data/_table_schema/test_get_column_type.py
