+++ Runtime/safe-ds/tests/data/_table_schema/test_table_equals.py
